refactor: 메인 로직을 controller로 분리
